Added skip if the example is not found

+++ docker-entrypoint.sh
@@ -17,7 +17,9 @@ link_config(){
 
 diff_config(){
   local basename=`basename $1`
-  diff /home/git/data/config/example/$basename.example $1.example | patch -N $1 || exit 1
+  if [ -e /home/git/data/config/example/$basename.example ];then
+    diff /home/git/data/config/example/$basename.example $1.example | patch -N $1 || exit 1
+  fi
   cp -pf $1.example /home/git/data/config/example
 }
